update state and action
